Refactoring, moving things around; start implementing axis range collecting needed for trim-var-glyphs
