+++ src/fontra/core/instancer.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 import logging
+from collections import defaultdict
 from dataclasses import dataclass, replace
 from enum import Enum
 from functools import cached_property, singledispatch
@@ -29,9 +30,11 @@ from .lrucache import LRUCache
 from .path import InterpolationError, PackedPath, joinPaths
 from .protocols import ReadableFontBackend
 from .varutils import (
+    AxisRange,
     locationToTuple,
     makeDenseLocation,
     mapAxesFromUserSpaceToSourceSpace,
+    subsetLocationKeep,
 )
 
 logger = logging.getLogger(__name__)
@@ -56,12 +59,23 @@ class FontInstancer:
         self._fontAxes: list[FontAxis | DiscreteFontAxis] | None = None
         self._fontSources: dict[str, FontSource] | None = None
         self._glyphErrors: set[str] = set()
+        self.variableGlyphAxisRanges: dict[str, dict[str, AxisRange]] | None = None
 
     async def _ensureSetup(self):
         if self._fontAxes is None:
             self._fontAxes = (await self.backend.getAxes()).axes
             assert self._fontSources is None
             self._fontSources = await self.backend.getSources()
+
+    def collectVariableGlyphAxisRanges(self):
+        self.variableGlyphAxisRanges = defaultdict(lambda: defaultdict(AxisRange))
+
+    def updateVariableGlyphAxisRanges(self, glyphName, location):
+        if self.variableGlyphAxisRanges is None:
+            return
+        glyphAxisRanges = self.variableGlyphAxisRanges[glyphName]
+        for axisName, value in location.items():
+            glyphAxisRanges[axisName].update(value)
 
     @cached_property
     def fontAxes(self) -> list[FontAxis | DiscreteFontAxis]:
@@ -251,17 +265,16 @@ class GlyphInstancer:
         )
         return instance
 
-    @cached_property
-    def fontAxisNames(self) -> set[str]:
-        return self.fontInstancer.fontAxisNames - {
-            axis.name for axis in self.glyph.axes
-        }
-
     def instantiate(
         self, location, *, coordSystem=LocationCoordinateSystem.SOURCE
     ) -> GlyphInstance:
         if coordSystem == LocationCoordinateSystem.USER:
             location = mapLocationFromUserToSource(location, self.fontAxes)
+
+        if self.fontInstancer.variableGlyphAxisRanges is not None:
+            self.fontInstancer.updateVariableGlyphAxisRanges(
+                subsetLocationKeep(location, self.glyphAxisNames)
+            )
 
         try:
             result = self.model.interpolateFromDeltas(location, self.deltas)
@@ -300,10 +313,6 @@ class GlyphInstancer:
             parentLocation,
             self.fontInstancer,
         )
-
-    @cached_property
-    def fontAxes(self) -> list[FontAxis | DiscreteFontAxis]:
-        return self.fontInstancer.fontAxes
 
     @cached_property
     def defaultFontSourceLocation(self) -> dict[str, float]:
@@ -355,8 +364,22 @@ class GlyphInstancer:
         ]
 
     @cached_property
+    def fontAxes(self) -> list[FontAxis | DiscreteFontAxis]:
+        return self.fontInstancer.fontAxes
+
+    @cached_property
+    def fontAxisNames(self) -> set[str]:
+        return self.fontInstancer.fontAxisNames - {
+            axis.name for axis in self.glyph.axes
+        }
+
+    @cached_property
+    def glyphAxisNames(self) -> set[str]:
+        return {axis.name for axis in self.glyph.axes}
+
+    @cached_property
     def combinedAxes(self) -> list[FontAxis | DiscreteFontAxis | GlyphAxis]:
-        glyphAxisNames = {axis.name for axis in self.glyph.axes}
+        glyphAxisNames = self.glyphAxisNames
         fontAxes = [axis for axis in self.fontAxes if axis.name not in glyphAxisNames]
         fontAxes = mapAxesFromUserSpaceToSourceSpace(fontAxes)
         return self.glyph.axes + fontAxes
